refactor(accounts): mv getTransactions() into file
We create a separate file for the getAccountTransactions() function so
that it may be more easily imported and shared between reports.  The
account report and multiple accounts report have been updated to reflect
this new location.
